Move statCache struct to internal/cache/metadata
